LayoutTests/inspector/heap/gc.html

<!doctype html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script>
function test()
{
    let suite = InspectorTest.createAsyncSuite("Heap.gc");

    suite.addTestCase({
        name: "TriggerGCShouldCreateGCEvent",
        description: "Calling Heap.gc should trigger Heap.garbageCollected event if enabled.",
        test: (resolve, reject) => {
            HeapAgent.gc();
            WebInspector.heapManager.singleFireEventListener(WebInspector.HeapManager.Event.GarbageCollected, (event) => {
                InspectorTest.expectThat(event.data.collection instanceof WebInspector.GarbageCollection, "Event should have GarbageCollection data.");
                InspectorTest.expectThat(event.data.collection.type === WebInspector.GarbageCollection.Type.Full, "GarbageCollection type should be Full.");
                resolve();
            });
        }
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Test for the Heap.gc command.</p>
</body>
</html>
