LayoutTests/inspector/heap/gc-expected.txt

Test for the Heap.gc command.


== Running test suite: Heap.gc
-- Running test case: TriggerGCShouldCreateGCEvent
PASS: Event should have GarbageCollection data.
PASS: GarbageCollection type should be Full.
